apps/backend/requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15,<3.17
channels>=4.1,<4.3
channels-redis>=4.2,<4.4
celery>=5.4,<5.6
django-cors-headers>=4.6,<4.8
gunicorn>=23.0,<24.0
psycopg[binary]>=3.2,<3.4
PyNaCl>=1.5,<2.0
boto3>=1.35,<2.0
redis>=5.1,<7.0
uvicorn[standard]>=0.34,<0.39
whitenoise>=6.8,<7.0
